requirements.txt
# Core Web Frameworks
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
nicegui>=2.0.0

# HTTP & API Communication
requests>=2.31.0
httpx>=0.27.0

# Environment Management
python-dotenv>=1.0.1

# Data Handling & Storage
pandas>=2.2.2
pydantic>=2.7.0

# Notion API Integration (future)
notion-client>=2.2.1

# Google Maps / Places API (future)
googlemaps>=4.10.0

# Async Tools & Utilities
aiofiles>=23.2.1
anyio>=4.3.0

# Developer Quality of Life
black>=24.4.2
isort>=5.13.2
flake8>=7.0.0
